[AI-131] Add videos endpoint
For now, we check that the user has the UserSocialAuth object
(logged in through SSO earlier). In the future, we are planning to add
new middleware to check a specific cookie containing
UserSocialAuth id. This cookie will be set in the social:complete view.

Added VIDEO_SIGNING_SERVICE_URL setting/site configuration to
be able to setup it in future.
Added SSO_PROVIDER_FOR_LOGIN_REDIRECT setting/site
configuration to be able to choose the provider for login redirect if
the user has no SSO login. **It should be equal to backend name
provided in admin/third_party_auth/oauth2providerconfig

Youtrack:
- https://youtrack.raccoongang.com/issue/AI-131